Feat: Renaming Entities / Add Component / Logger + Viewspace
Added renaming functionality for entities, as well as the ability to enable and disable.

Added a button to add components to entity (only transform component exists as of writing).

Created a logger so all errors can be read inside the engine instead of on VS also added a view for the logger.

BUG FIXES:
Fixed a bug that caused a crash when nothing was selected upon hitting Undo/Redo

Fixed a bug that allowed calling the UndoRedo.Add function twice in the same call.

Fixed a bug that caused newly added entities to not get default transform component.

Fixed a bug causing undoredo list to stay populated from another project. Not unloading correctly.